pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gnn-signal-optimizer"
version = "0.1.0"
description = "GNN-based traffic signal optimizer for SUMO"
authors = [{name = "OpenAI", email = "example@example.com"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.0.0",
    "numpy>=1.24",
    "pandas>=2.0",
    "scipy>=1.10",
    "tqdm>=4.65",
    "loguru>=0.7",
    "pydantic>=2.5",
    "fastapi>=0.103",
    "uvicorn>=0.23",
    "websockets>=11.0",
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    "prometheus-client>=0.16",
    "pyyaml>=6.0",
    "python-socketio>=5.8",
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
    "httpx>=0.25",
    "pyarrow>=12.0",
    "networkx>=3.1",
    "torch_geometric>=2.4",
    "sumolib>=1.18.0",
    "traci>=1.18.0"
]

[project.optional-dependencies]
dev = [
    "pytest-cov>=4.1",
    "black>=23.7",
    "ruff>=0.1.3"
]

[tool.pytest.ini_options]
addopts = "-ra"
asyncio_mode = "auto"

[tool.setuptools.packages.find]
where = ["."]
include = ["gso*"]
